requirements.txt
python-dotenv
disnake